Documentation

All docstrings!

AstrodynamicalModels.AstrodynamicalModels — Module

Provides common astrodynamical models through ModelingToolkit, physical attributes for solar system bodies through SPICE, and ephemeris model downloading and interpolation through HORIZONS and Interpolations!

Tip

Check out the ModelingToolkit docs to learn how to use these systems for orbit propagation with DifferentialEquations, or see GeneralAstrodynamics for some convenient orbit propagation wrappers.

Usage

Physical Attributes

naifcode("earth") == 399
naifname("mars") == 499

μₑ = massparameter("earth")
Rₘ = meanradius("mars")

Dynamics

using ModelingToolkit

length(states(R2BP())) == 6
length(states(NBP(2))) == 12

Ephemeris

Warning

Before usage, see the warning about copyrighted tools in this project's README, or in the Extended Help section of this docstring!

using HORIZONS
ephemeris("earth") isa DataFrame

using Dates
ephemeris("earth", (now()-Year(5), now()+Year(5)), Day(1)) isa DataFrame

using Interpolations: CubicHermite
ContinuousEphemeris("earth") isa ContinuousEphemeris{<:Number, <:Number, <:CubicHermite, <:CubicHermite, <:CubicHermite}

Extended help

License

Warning

The license in this docstring applies to all code in this package, but note that the ephemeris data downloading functionality of this package (e.g. the ephemeris function) uses copyrighted tools with their own usage and code sharing restrictions. These copyrighted tools are owned by the Jet Propulsion Laboratory at the California Institute of Technology. These copyrighted tools are, in part, located at ftp://ssd.jpl.nasa.gov/pub/ssd/SCRIPTS/.

Open source wrappers for these copyrighted tools exist, e.g. HORIZONS.jl and AstroPy's astroquery. The ephemeris handling functionality of this package uses HORIZONS.jl.

It is the responsibility of the user to verify they meet the necessary requirements, as specifed by JPL in their scripts (e.g. vec_tbl), before they share or use the copyrighted tools. This includes the ephemeris handling functionality in this package (e.g. ephemeris).

MIT License

Copyright (c) 2021 Joe Carpinelli

Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

Exports

Always

naifcode

naifname

axialradii

meanradius

massparameter

Requires ModelingToolkit

NBP

R2BP

CR3BP

R2BPFunction

CR3BPFunction

NBPFunction

Requires HORIZONS

ephemeris

Requires Interpolations (and HORIZONS)

ContinuousEphemeris

Imports

Always

Requires, SPICE, LinearAlgebra, DocStringExtensions

Requires ModelingToolkit

Memoize

Requires HORIZONS

CSV, DataFrames

AstrodynamicalModels.AbstractEphemerisInterpolation — Type

abstract type AbstractEphemerisInterpolation

An abstract supertype for all possible interpolations for ephemeris data.

AstrodynamicalModels.ContinuousEphemeris — Type

A callable structure which holds ephemeris data for a body's positions and velocities between the timepoints specified by T. Call the object with a Julian day input to interpolate the body's positions and velocities using Interpolations.CubicHermite.

AstrodynamicalModels.ContinuousEphemeris — Method

Construct a ContinuousEphemeris instance from an appropriately formatted DataFrame. This is mostly an internal package method. See ephemeris(args...; continuous = Val{true}, kwargs...) for a more user friendly constructor!

AstrodynamicalModels.ContinuousEphemeris — Method

Interpolate the body's position and velocity by a Julian day input.

AstrodynamicalModels.CR3BP — Method

CR3BP(; stm, structural_simplify, name)

A ModelingToolkit.ODESystem for the Circular Restricted Three-body Problem.

The order of the states follows: [x, y, z, ẋ, ẏ, ż].

The order of the parameters follows: [μ].

Extended Help

The Circular Restricted Three-body Problem is a simplified dynamical model describing one small body (spacecraft, etc.) and two celestial bodies moving in a circle about their common center of mass. This may seem like an arbitrary simplification, but this assumption holds reasonably well for the Earth-Moon, Sun-Earth, and many other systems in our solar system.

Usage

model = CR3BP(; stm=true) 

AstrodynamicalModels.CR3BPFunction — Method

CR3BPFunction(; stm, structural_simplify, name, kwargs...)

Returns an ODEFunction for CR3BP dynamics. Results are cached with Memoize.jl.

The order of the states follows: [x, y, z, ẋ, ẏ, ż].

The order of the parameters follows: [μ].

Extended Help

Usage

The stm, structural_simplify, and name keyword arguments are passed to CR3BP. All other keyword arguments are passed directly to SciMLBase.ODEFunction.

f = CR3BPFunction(; stm=false, jac=true)
let u = randn(6), p = randn(1), t = 0
    f(u, p, t)
end

AstrodynamicalModels.NBP — Method

NBP(N; stm, structural_simplify, name)

A ModelingToolkit.ODESystem for the Newtonian N-body Problem.

The order of the states follows: [x₁, y₁, z₁, ..., xₙ, yₙ, zₙ, ẋ₁, ẏ₁, ż₁, ..., ẋₙ, ẏₙ, żₙ].

The order of the parameters follows: [G, m₁, m₂, ..., mₙ].

Warning

Be careful about specifying stm=true for systems with N ≥ 3! If state transition matrix dynamics are enabled, you can calculate the total number of system states with N*6 + (N*6)^2. Note that this increases exponentially as N grows! For N == 6, unless you're using parallelization, your computer may run for several hours.

Extended Help

The N-body problem is a model which describes how N bodies will move with respect to a common origin. This problem typically involves many bodies which act due to one force: electromagentism, gravity, etc. This model applies most closely to many celestial bodies moving due to gravity. That's about right for a model in a package called AstrodynamicalModels!

Usage

# One model for ALL the planets in our solar system 😎
model = NBP(9) 

AstrodynamicalModels.NBPFunction — Method

NBPFunction(N; stm, structural_simplify, name, kwargs...)

Returns an ODEFunction for NBP dynamics. Results are cached with Memoize.jl. The order of states and parameters in the ODEFunction arguments are equivalent to the order of states and parameters for the system produced with NBP(N). As a general rule, the order of the states follows: [x₁, y₁, z₁, ..., xₙ, yₙ, zₙ, ẋ₁, ẏ₁, ż₁, ..., ẋₙ, ẏₙ, żₙ].

Note

Unlike R2BP and CR3BP, jac is set to false by default. The number of states for NBP systems can be very large for relatively small numbers of bodies (N). Enabling jac=true by default would cause unnecessarily long waiting times for this function to return for N ≥ 3 or so. If N=2 and stm=true, setting jac=true could still result in several minutes of calculations, depending on the computer you're using.

Warning

Be careful about specifying stm=true for systems with N ≥ 3! If state transition matrix dynamics are enabled, you can calculate the total number of system states with N*6 + (N*6)^2. Note that this increases exponentially as N grows! For N == 6, unless you're using parallelization, your computer may run for several hours.

Extended Help

Usage

The stm, structural_simplify, and name keyword arguments are passed to NBP. All other keyword arguments are passed directly to SciMLBase.ODEFunction.

f = NBPFunction(3; stm=false, structural_simplify=true, name=:NBP, jac=false, sparse=false)
let u = randn(3*6), p = randn(1 + 3), t = 0
    f(u, p, t)
end

AstrodynamicalModels.R2BP — Method

R2BP(; stm, structural_simplify, name)

A ModelingToolkit.ODESystem for the Restricted Two-body Problem.

The order of the states follows: [x, y, z, ẋ, ẏ, ż].

The order of the parameters follows: [μ].

Extended Help

The Restricted Two-body Problem is a simplified dynamical model describing one small body (spacecraft, etc.) and one celestial body. The gravity of the celestial body exhibits a force on the small body. This model is commonly used as a simplification to descibe our solar systems' planets orbiting our sun, or a spacecraft orbiting Earth.

Usage

model = R2BP() 

AstrodynamicalModels.R2BPFunction — Method

R2BPFunction(; stm, structural_simplify, name, kwargs...)

Returns an ODEFunction for R2BP dynamics. Results are cached with Memoize.jl.

The order of the states follows: [x, y, z, ẋ, ẏ, ż].

The order of the parameters follows: [μ].

Extended Help

Usage

The stm, structural_simplify, and name keyword arguments are passed to R2BP. All other keyword arguments are passed directly to SciMLBase.ODEFunction.

f = R2BPFunction(; stm=false, structural_simplify=true, name=:R2BP, jac=true)
let u = randn(6), p = randn(1), t = 0
    f(u, p, t)
end

AstrodynamicalModels.ephemeris — Function

ephemeris(body)
ephemeris(body, timespan)
ephemeris(body, timespan, intervol; continuous, type, email, wrt, epoch)

Given a celestial body's name, or the NAIF ID of the body, return a DataFrame of ephemeris data for the specified timespan and intervol between time steps.

Warning

The license file for this package applies to all code in this package, but note that the ephemeris data handling functionality of this package (e.g. the ephemeris function) uses copyrighted tools with their own usage and code sharing restrictions. These copyrighted tools are owned by the Jet Propulsion Laboratory at the California Institute of Technology. These copyrighted tools are, in part, located at ftp://ssd.jpl.nasa.gov/pub/ssd/SCRIPTS/.

Open source wrappers for these copyrighted tools exist, e.g. HORIZONS.jl and AstroPy's astroquery. The ephemeris handling functionality of this package uses HORIZONS.jl.

It is the responsibility of the user to verify they meet the necessary requirements, as specifed by JPL in their copyrighted tools (e.g. vec_tbl), before they share or use those tools.

AstrodynamicalModels.ephemeris — Method

ephemeris(body, time; kwargs...)

Returns the Cartesian state vector of the requested body at the provided date and time! All kwargs are passed directly to the timespan method for ephemeris.

Warning

The license file for this package applies to all code in this package, but note that the ephemeris data handling functionality of this package (e.g. the ephemeris function) uses copyrighted tools with their own usage and code sharing restrictions. These copyrighted tools are owned by the Jet Propulsion Laboratory at the California Institute of Technology. These copyrighted tools are, in part, located at ftp://ssd.jpl.nasa.gov/pub/ssd/SCRIPTS/.

Open source wrappers for these copyrighted tools exist, e.g. HORIZONS.jl and AstroPy's astroquery. The ephemeris handling functionality of this package uses HORIZONS.jl.

It is the responsibility of the user to verify they meet the necessary requirements, as specifed by JPL in their copyrighted tools (e.g. vec_tbl), before they share or use those tools.

AstrodynamicalModels.naifcode — Method

naifcode(name)

Return the NAIF ID for the provided celestial body name. If no ID for the provided body is found, a KeyError exception is thrown.

Note

This is a simple wrapper around SPICE.bodn2c, which is itself a wrapper around the CSPICE library's bod2nc function! This integer code lookup is robust to different capitalizations. For example, "Earth", "earth", and "eArTH" should all return integer code 399.

AstrodynamicalModels.naifname — Method

naifname(code)

Return the celestial body name for the provided NANIF ID. If no name for the provided ID is found, a KeyError exception is thrown.

Note

This is a simple wrapper around SPICE.bodc2n, which is itself a wrapper around the CSPICE library's bodc2n function!