Documentation

All docstrings!

AstrodynamicalModels.AstrodynamicalModels — Module

Provides astrodynamical models as ModelingToolkit.ODESystems. Check out the ModelingToolkit docs to learn how to use these systems for orbit propagation with DifferentialEquations, or see GeneralAstrodynamics for some convenient orbit propagation wrappers.

Extended help

Exports

CR3BP

CR3BPFunction

R2BP

R2BPFunction

Imports

Base

Core

DocStringExtensions

LinearAlgebra

ModelingToolkit

StaticArrays

Symbolics

AstrodynamicalModels.CR3BP — Method

CR3BP(; stm, structural_simplify, name)

A ModelingToolkit.ODESystem for the Circular Restricted Three-body Problem.

Extended Help

The Circular Restricted Three-body Problem is a simplified dynamical model describing one small body (spacecraft, etc.) and two celestial bodies moving in a circle about their common center of mass. This may seem like an arbitrary simplification, but this assumption holds reasonably well for the Earth-Moon, Sun-Earth, and many other systems in our solar system.

AstrodynamicalModels.CR3BPFunction — Method

CR3BPFunction(; stm, structural_simplify, name, kwargs...)

Returns an ODEFunction for CR3BP dynamics. Results are cached with Memoize.jl.

Extended Help

Usage

The stm, structural_simplify, and name keyword arguments are passed to CR3BP. All other keyword arguments are passed directly to SciMLBase.ODEFunction.

f = CR3BPFunction(; stm=false, jac=true)

AstrodynamicalModels.R2BP — Method

R2BP(; stm, structural_simplify, name)

A ModelingToolkit.ODESystem for the Restricted Two-body Problem.

Extended Help

The Restricted Two-body Problem is a simplified dynamical model describing one small body (spacecraft, etc.) and one celestial body. The gravity of the celestial body exhibits a force on the small body. This model is commonly used as a simplification to descibe our solar systems' planets orbiting our sun, or a spacecraft orbiting Earth.

AstrodynamicalModels.R2BPFunction — Method

R2BPFunction(; stm, structural_simplify, name, kwargs...)

Returns an ODEFunction for R2BP dynamics. Results are cached with Memoize.jl.

Extended Help

Usage

The stm, structural_simplify, and name keyword arguments are passed to R2BP. All other keyword arguments are passed directly to SciMLBase.ODEFunction.

f = R2BPFunction(; stm=false, structural_simplify=true, name=:R2BP, jac=true)